chore: move UsersFilter to components
So it can be reused by the group members page.
