+++ site/src/pages/UsersPage/UsersPageView.tsx
@@ -1,6 +1,7 @@
 import type { GroupsByUserId } from "api/queries/groups";
 import type * as TypesGen from "api/typesGenerated";
 import { Button } from "components/Button/Button";
+import { UsersFilter } from "components/Filter/UsersFilter";
 import {
 	PaginationContainer,
 	type PaginationResult,
@@ -13,7 +14,6 @@ import {
 import { UserPlusIcon } from "lucide-react";
 import type { ComponentProps, FC } from "react";
 import { Link as RouterLink } from "react-router";
-import { UsersFilter } from "./UsersFilter";
 import { UsersTable } from "./UsersTable/UsersTable";
 
 interface UsersPageViewProps {
